ytarasovych: remove methods isAuthorizationButtonVisible, isRegistrationButtonVisible, isLocationSelectionVisible, isHelpCenterButtonVisible and isContactUsButtonVisible from Header to MobileMenu
