+++ src/main/java/com/softserveinc/ita/rozetka/components/Header.java
@@ -33,26 +33,6 @@ public class Header {
         return isVisible("//rz-user-identification");
     }
 
-    public boolean isAuthorizationButtonVisible() {
-        return isVisible("(" + authorizationXpath + ")[1]");
-    }
-
-    public boolean isRegistrationButtonVisible() {
-        return isVisible("(" + authorizationXpath + ")[2]");
-    }
-
-    public boolean isLocationSelectionVisible() {
-        return isVisible("//button[contains(@class,'city-toggle')]");
-    }
-
-    public boolean isHelpCenterButtonVisible() {
-        return isVisible("//a[contains(@class,'side-menu__button') and contains(@href,'help.')]");
-    }
-
-    public boolean isContactUsButtonVisible() {
-        return isVisible("//a[contains(@class,'side-menu__button') and contains(@href,'t.me')]");
-    }
-
     public boolean isShoppingCartEmpty() {
         return isVisible("//button[@rzopencart='']//span[contains(@class, 'counter')]");
     }
